move guideline/classification functions to errorlogger
