+++ cli/cppcheckexecutor.cpp
@@ -23,7 +23,6 @@
 #include "cppcheckexecutor.h"
 
 #include "analyzerinfo.h"
-#include "checkers.h"
 #include "checkersreport.h"
 #include "cmdlinelogger.h"
 #include "cmdlineparser.h"
@@ -628,9 +627,9 @@ void StdLogger::reportErr(const ErrorMessage &msg)
         return;
 
     ErrorMessage msgCopy = msg;
-    msgCopy.guideline = checkers::getGuideline(msgCopy.id, mSettings.reportType,
-                                               mSettings.guidelineMapping, msgCopy.severity);
-    msgCopy.classification = checkers::getClassification(msgCopy.guideline, mSettings.reportType);
+    msgCopy.guideline = getGuideline(msgCopy.id, mSettings.reportType,
+                                     mSettings.guidelineMapping, msgCopy.severity);
+    msgCopy.classification = getClassification(msgCopy.guideline, mSettings.reportType);
 
     if (mSettings.outputFormat == Settings::OutputFormat::sarif)
         mSarifReport.addFinding(msgCopy);
