+++ lib/checkers.cpp
@@ -2097,183 +2097,3 @@ std::vector<checkers::Info> checkers::certCppInfo{
     {"MSC53-CPP", "L2"},
     {"MSC54-CPP", "L2"},
 };
-
-namespace checkers {
-    std::string getClassification(const std::string &guideline, ReportType reportType) {
-        if (guideline.empty())
-            return "";
-
-        const auto getClassification = [](const std::vector<checkers::Info> &info, const std::string &guideline) -> std::string {
-            const auto it = std::find_if(info.cbegin(), info.cend(), [&](const checkers::Info &i) {
-                return caseInsensitiveStringCompare(i.guideline, guideline) == 0;
-            });
-            if (it == info.cend())
-                return "";
-            return it->classification;
-        };
-
-        switch (reportType) {
-        case ReportType::autosar:
-            return getClassification(checkers::autosarInfo, guideline);
-        case ReportType::certC:
-            return getClassification(checkers::certCInfo, guideline);
-        case ReportType::certCpp:
-            return getClassification(checkers::certCppInfo, guideline);
-        case ReportType::misraC:
-        {
-            auto components = splitString<std::vector>(guideline, '.');
-            if (components.size() != 2)
-                return "";
-
-            const int a = std::stoi(components[0]);
-            const int b = std::stoi(components[1]);
-
-            const std::vector<checkers::MisraInfo> &info = checkers::misraC2012Rules;
-            const auto it = std::find_if(info.cbegin(), info.cend(), [&](const checkers::MisraInfo &i) {
-                    return i.a == a && i.b == b;
-                });
-
-            if (it == info.cend())
-                return "";
-
-            return it->str;
-        }
-        case ReportType::misraCpp2008:
-        case ReportType::misraCpp2023:
-        {
-            char delim;
-            const std::vector<checkers::MisraCppInfo> *info;
-            if (reportType == ReportType::misraCpp2008) {
-                delim = '-';
-                info = &checkers::misraCpp2008Rules;
-            } else {
-                delim = '.';
-                info = &checkers::misraCpp2023Rules;
-            }
-
-            auto components = splitString<std::vector>(guideline, delim);
-            if (components.size() != 3)
-                return "";
-
-            const int a = std::stoi(components[0]);
-            const int b = std::stoi(components[1]);
-            const int c = std::stoi(components[2]);
-
-            const auto it = std::find_if(info->cbegin(), info->cend(), [&](const checkers::MisraCppInfo &i) {
-                    return i.a == a && i.b == b && i.c == c;
-                });
-
-            if (it == info->cend())
-                return "";
-
-            return it->classification;
-        }
-        default:
-            return "";
-        }
-    }
-
-    std::string getGuideline(const std::string &errId, ReportType reportType,
-                             const std::map<std::string, std::string> &guidelineMapping,
-                             Severity severity)
-    {
-        std::string guideline;
-
-        switch (reportType) {
-        case ReportType::autosar:
-            if (errId.rfind("premium-autosar-", 0) == 0) {
-                guideline = errId.substr(16);
-                break;
-            }
-            if (errId.rfind("premium-misra-cpp-2008-", 0) == 0)
-                guideline = "M" + errId.substr(23);
-            break;
-        case ReportType::certC:
-        case ReportType::certCpp:
-            if (errId.rfind("premium-cert-", 0) == 0) {
-                guideline = errId.substr(13);
-                std::transform(guideline.begin(), guideline.end(),
-                               guideline.begin(), static_cast<int (*)(int)>(std::toupper));
-            }
-            break;
-        case ReportType::misraC:
-            if (errId.rfind("misra-c20", 0) == 0)
-                guideline = errId.substr(errId.rfind('-') + 1);
-            break;
-        case ReportType::misraCpp2008:
-            if (errId.rfind("misra-cpp-2008-", 0) == 0)
-                guideline = errId.substr(15);
-            break;
-        case ReportType::misraCpp2023:
-            if (errId.rfind("misra-cpp-2023-", 0) == 0)
-                guideline = errId.substr(15);
-            break;
-        default:
-            break;
-        }
-
-        if (!guideline.empty())
-            return guideline;
-
-        auto it = guidelineMapping.find(errId);
-
-        if (it != guidelineMapping.cend())
-            return it->second;
-
-        if (severity != Severity::error && severity != Severity::warning)
-            return "";
-
-        it = guidelineMapping.find(errId);
-
-        if (it != guidelineMapping.cend())
-            return it->second;
-
-        return "";
-    }
-
-    std::map<std::string, std::string> createGuidelineMapping(ReportType reportType) {
-        std::map<std::string, std::string> guidelineMapping;
-        const std::vector<IdMapping> *idMapping1 = nullptr;
-        const std::vector<IdMapping> *idMapping2 = nullptr;
-        std::string ext1, ext2;
-
-        switch (reportType) {
-        case ReportType::autosar:
-            idMapping1 = &idMappingAutosar;
-            break;
-        case ReportType::certCpp:
-            idMapping2 = &idMappingCertCpp;
-            ext2 = "-CPP";
-            FALLTHROUGH;
-        case ReportType::certC:
-            idMapping1 = &idMappingCertC;
-            ext1 = "-C";
-            break;
-        case ReportType::misraC:
-            idMapping1 = &idMappingMisraC;
-            break;
-        case ReportType::misraCpp2008:
-            idMapping1 = &idMappingMisraCpp2008;
-            break;
-        case ReportType::misraCpp2023:
-            idMapping1 = &idMappingMisraCpp2023;
-            break;
-        default:
-            break;
-        }
-
-        if (idMapping1) {
-            for (const auto &i : *idMapping1)
-                for (const std::string &cppcheckId : splitString<std::vector>(i.cppcheckId, ','))
-                    guidelineMapping[cppcheckId] = i.guideline + ext1;
-        }
-
-        if (idMapping2) {
-            for (const auto &i : *idMapping2)
-                for (const std::string &cppcheckId : splitString<std::vector>(i.cppcheckId, ','))
-                    guidelineMapping[cppcheckId] = i.guideline + ext2;
-        }
-
-        return guidelineMapping;
-    }
-}
